Update document-store-example.md
Updated examples, the last one has changed a bit. In the actual docs page, this is using a different random collection. From my point of view if the example is within a site following a structured or sequential order of operations I think changing the collection’s name and going back 4 or 5 steps back to reproduce the same operations is not very instructive. I think it’s better to continue with operations following the same schema that the reader has been following, in this case with the employees collection.